break circular dependency
